LayoutTests/fast/js/dfg-array-dead-expected.txt

Tests that an array being dead does not result in register allocation failures.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo() is 2 on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
